pyproject.toml
[project]
name = "madr-postgres"
version = "0.1.0"
description = ""
authors = [
    {name = "devisono",email = "devison_santos@yahoo.com.br"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard] (>=0.124.4,<0.125.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.45,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev madr_postgres/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=madr_postgres -vv'
post_test = 'coverage html'

[tool.coverage.run]
core = 'ctrace'
concurrency = ['thread', 'greenlet']
[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.9,<0.15.0)",
    "testcontainers (>=4.13.3,<5.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)"
]
